+++ src/chains/details/mainnet.ts
@@ -1,5 +1,5 @@
 import { baseNativeCurrency } from 'src/common/consts/tokens'
-import { ChainInfo, SupportedChainId } from '../../common/types/chains'
+import { ChainInfo, SupportedChainId } from '../types'
 import { RAW_CHAINS_FILES_PATH } from '../const'
 
 export const ethereumLogo = `${RAW_CHAINS_FILES_PATH}/images/mainnet-logo.svg`
