+++ src/index.ts
@@ -5,4 +5,4 @@ export * from './order-signing'
 export * from './composable'
 export * from './cow-shed'
 export * from './trading'
-export * from './chains/const'
+export * from './chains'
